move Version from status to settings
