+++ interface/src/AuthenticatedRouting.tsx
@@ -16,6 +16,7 @@ import DownloadUpload from 'app/settings/DownloadUpload';
 import MqttSettings from 'app/settings/MqttSettings';
 import NTPSettings from 'app/settings/NTPSettings';
 import Settings from 'app/settings/Settings';
+import Version from 'app/settings/Version';
 import Network from 'app/settings/network/Network';
 import Security from 'app/settings/security/Security';
 import APStatus from 'app/status/APStatus';
@@ -26,7 +27,6 @@ import NTPStatus from 'app/status/NTPStatus';
 import NetworkStatus from 'app/status/NetworkStatus';
 import Status from 'app/status/Status';
 import SystemLog from 'app/status/SystemLog';
-import Version from 'app/status/Version';
 import { Layout } from 'components';
 import { AuthenticatedContext } from 'contexts/authentication';
 
@@ -49,11 +49,11 @@ const AuthenticatedRouting = memo(() => {
         <Route path="/status/ntp" element={<NTPStatus />} />
         <Route path="/status/ap" element={<APStatus />} />
         <Route path="/status/network" element={<NetworkStatus />} />
-        <Route path="/status/version" element={<Version />} />
 
         {me.admin && (
           <>
             <Route path="/settings" element={<Settings />} />
+            <Route path="/settings/version" element={<Version />} />
             <Route path="/settings/application" element={<ApplicationSettings />} />
             <Route path="/settings/mqtt" element={<MqttSettings />} />
             <Route path="/settings/ntp" element={<NTPSettings />} />
